Add DbConnectionFactory
